Refund step moved to start
When applying for a refund we need to know the applicants
income/benefits _at the time they paid the fee_.

Moving the refund step to the start of the form allows us to rephrase
the income and benefit questions for this scenario.

It also allows for a better flow rather than starting with their
relationship status.