Revised digital paper flow (Pt 2)
Focus on the last two pages of the flow.
Created a confirmation-paper.slim page and a confirmation-post.slim
page.
The button copy on the summary.slim page has been changed
Each of the confirmation pages reference the form name that the user
types in Step 1. If the use doesn’t know their form, it will default to
claim form. (Update to come via Clive)